Remove public sign-up routes and links
